requirements.txt
flask==1.0.2
Bootstrap-Flask==1.0.6
requests==2.19.1
python-dotenv==0.9.1
